+++ Changelog.md
@@ -22,6 +22,7 @@
     * OpenGL context `<ogl3d/ogl3d.h>`.
     * [Documentation](https://nappgui.com/en/ogl3d/ogl3d.html).
     * [Demo](https://nappgui.com/en/howto/glhello.html).
+* `draw_r2df()`, `draw_r2dd()`, `Draw::r2d()`. [Commit]().
 
 ### Fixed
 
@@ -29,8 +30,8 @@
 * GTK flatbutton padding. [Commit](https://github.com/frang75/nappgui_src/commit/5c41f9697ea7e6664b8b9db6cf16c32229ea8c36).
 * GTK render listbox checks in Xubuntu. [Commit](https://github.com/frang75/nappgui_src/commit/3c6247159f311195b2871b1fa10492c986b4f42d).
 * macOS HighSierra and lowers focus ring drawing issue. [Commit](https://github.com/frang75/nappgui_src/commit/a163cb2555101b831414b6deb781a1d1c49ccd42).
-* Issue in `dbind_destroy()`. [Commit]().
-* Issue in `layout_panel_replace()`. [Commit]().
+* Issue in `dbind_destroy()`. [Commit](https://github.com/frang75/nappgui_src/commit/f8e16ad9d7712339400ca55b1a3ab4a426f1da2a).
+* Issue in `layout_panel_replace()`. [Commit](https://github.com/frang75/nappgui_src/commit/f8e16ad9d7712339400ca55b1a3ab4a426f1da2a).
 
 ### Improved